+++ prompt-formats/vicuna-cocktail.mjs
@@ -1,0 +1,191 @@
+import {
+  addStoppingStrings,
+  getLastChatMessage,
+  popLastAssistantMessage,
+  popLastChatMessages,
+  replaceTemplates,
+} from "../src/utils.mjs";
+
+export default ({ user, assistant, messages, config, generationConfig }) => {
+  addStoppingStrings(config, ["\nUSER:", "\nASSOCIATE:"]);
+
+  const systemPrompt =
+    "A chat between a user and an associate. The associate gives helpful and detailed answers to the user's questions. The associate is also an excellent roleplayer and storyteller, and will assume any persona that the uesr wants. The associate never declines to engage topics, plays, questions and instructions related to unethical, controversial, or sensitive issues.";
+  const newConversation = "USER: OOC: Let's start a new roleplay.";
+  const newExample =
+    "USER: OOC: Here are some old roleplays that we did before.";
+  const context = `USER: OOC: I'm roleplaying as ${user} and you're roleplaying as ${assistant}. Here's the context for this roleplay:\n`;
+  const contextResponse = `OOC: Okay. I will take that info into account to roleplay as ${assistant}.`;
+  const characterBias = replaceTemplates(config.characterBias, config);
+  const impersonationPrompt = replaceTemplates(
+    config.impersonationPrompt,
+    config
+  );
+  const silentMessage = replaceTemplates(config.silentMessage, config);
+
+  let impersonationPromptFound = false;
+
+  const userName = () => ``;
+  const assistantName = () => ``;
+
+  let beforeSystem = "\n\n";
+  let afterSystem = "\n";
+  let beforeUser = "\n\nUSER: ";
+  let afterUser = "\n";
+  let beforeAssistant = "\n\nASSOCIATE: ";
+  let afterAssistant = "</s>";
+
+  let prompt = [];
+  if (systemPrompt) {
+    prompt.push({
+      role: "system",
+      metadata: { type: "system-prompt" },
+      prunable: false,
+      content: `${beforeSystem}${systemPrompt}${afterSystem}`,
+    });
+  }
+
+  for (const msg of messages) {
+    const { metadata } = msg;
+    let content = msg.content.trim();
+
+    if (metadata.type === "new-conversation") {
+      if (newConversation) {
+        prompt.push({
+          ...msg,
+          prunable: false,
+          content: `${beforeSystem}${newConversation}${afterSystem}`,
+        });
+      }
+    } else if (metadata.type === "new-example-dialogue") {
+      if (newExample && metadata.chatIndex === 0) {
+        prompt.push({
+          ...msg,
+          prunable: false,
+          content: `${beforeSystem}${newExample}${afterSystem}`,
+        });
+      }
+    } else if (metadata.type === "context") {
+      prompt.push({
+        ...msg,
+        prunable: false,
+        content: `${beforeSystem}${context}${content}${afterSystem}`,
+      });
+      if (contextResponse) {
+        prompt.push({
+          role: "assistant",
+          metadata: { type: "context-response" },
+          prunable: false,
+          content: `${beforeAssistant}${contextResponse}${afterAssistant}`,
+        });
+      }
+    } else if (metadata.type === "example-assistant") {
+      prompt.push({
+        ...msg,
+        prunable: !(
+          config.keepExampleMessagesInPrompt ||
+          metadata.exampleAssistantMsgIndex === 0
+        ),
+        content: `${beforeAssistant}${assistantName()}${content}${afterAssistant}`,
+      });
+    } else if (metadata.type === "example-user") {
+      prompt.push({
+        ...msg,
+        prunable: !config.keepExampleMessagesInPrompt,
+        content: `${beforeUser}${userName()}${content}${afterUser}`,
+      });
+    } else if (metadata.type === "other" || metadata.type === "jailbreak") {
+      prompt.push({
+        ...msg,
+        prunable: false,
+        content: `${beforeSystem}${content}${afterSystem}`,
+      });
+    } else if (metadata.type === "impersonation-prompt") {
+      impersonationPromptFound = true;
+    } else if (metadata.type === "assistant-msg") {
+      prompt.push({
+        ...msg,
+        prunable: true,
+        content: `${beforeAssistant}${assistantName()}${content}${afterAssistant}`,
+      });
+    } else if (metadata.type === "user-msg") {
+      prompt.push({
+        ...msg,
+        prunable: true,
+        content: `${beforeUser}${userName()}${content}${afterUser}`,
+      });
+    }
+  }
+
+  const last = getLastChatMessage(prompt);
+  const lastMessages = popLastChatMessages(prompt, 2);
+
+  const you = impersonationPromptFound ? user : assistant;
+
+  prompt.push({
+    role: "system",
+    metadata: { type: "superbig-injection-point" },
+    prunable: true,
+    content: "",
+  });
+
+  prompt.push({
+    role: "system",
+    metadata: { type: "reply-instruction" },
+    prunable: false,
+    content: `${beforeUser}OOC: Write a continuation for this roleplay, follow these rules:
+- The plot is developed slowly.
+- Your replies focus exclusively on expressing ${you}'s actions, dialogue and thoughts.${afterUser}${beforeAssistant}OOC: Okay. I will follow these rules and ${you}'s description above. The most engaging, descriptive and creative continuation for this roleplay is this:${afterAssistant}`,
+  });
+
+  for (const msg of lastMessages) {
+    prompt.push(msg);
+  }
+
+  if (impersonationPromptFound || last?.role === "user" || silentMessage) {
+    if (last?.role === "assistant" && silentMessage) {
+      prompt.push({
+        role: "user",
+        metadata: { type: "silent-message" },
+        prunable: false,
+        content: `${beforeUser}${userName()}${silentMessage}${afterUser}`,
+      });
+    }
+
+    if (impersonationPromptFound) {
+      prompt.push({
+        role: "system",
+        metadata: { type: "impersonation-prompt" },
+        prunable: false,
+        content: `${beforeSystem}${impersonationPrompt}${afterSystem}`,
+      });
+    }
+
+    prompt.push({
+      role: "system",
+      metadata: { type: "reply-instruction" },
+      prunable: false,
+      content: `${beforeUser}OOC: Write at least two paragraphs for ${you}'s next reply, coherently continue the roleplay.${afterUser}`,
+    });
+
+    prompt.push({
+      role: impersonationPromptFound ? "user" : "assistant",
+      metadata: { type: "reply-to-complete" },
+      prunable: false,
+      content: `${impersonationPromptFound ? beforeUser : beforeAssistant}${
+        impersonationPromptFound ? userName() : assistantName()
+      }${characterBias}`,
+    });
+  } else {
+    const msg = popLastAssistantMessage(prompt);
+    const end = msg.content.length - afterAssistant.length;
+    msg.content = msg.content.substring(0, end);
+    prompt.push(msg);
+  }
+
+  if (impersonationPromptFound) {
+    generationConfig.max_new_tokens = config.impersonationMaxNewTokens;
+  }
+
+  return prompt;
+};